Made firestore user dao and repository, using ViewModels for better performance of firestore actions.
