+++ app/src/main/java/com/android/carrent/firestore/FirestoreUserRepository.kt
@@ -1,0 +1,13 @@
+package com.android.carrent.firestore
+
+import com.android.carrent.utils.Constants
+import com.google.firebase.firestore.DocumentReference
+import com.google.firebase.firestore.FirebaseFirestore
+
+class FirestoreUserRepository: IFirestoreUserDao {
+    override fun getUser(uid: String): DocumentReference {
+        val ref = FirebaseFirestore.getInstance().collection(Constants.FIRESTORE_USERS_REFERENCE).document(uid)
+        return ref
+    }
+
+}
