+++ app/src/main/java/com/android/carrent/firestore/IFirestoreUserDao.kt
@@ -1,0 +1,7 @@
+package com.android.carrent.firestore
+
+import com.google.firebase.firestore.DocumentReference
+
+interface IFirestoreUserDao {
+    fun getUser(uid: String): DocumentReference
+}
