+++ app/src/main/java/com/android/carrent/fragments/main/RegisterFragment.kt
@@ -11,11 +11,10 @@ import android.view.ViewGroup
 
 import com.android.carrent.R
 import com.android.carrent.models.User
-import com.android.carrent.utils.*
 import com.android.carrent.utils.Constants.FIRESTORE_USERS_REFERENCE
+import com.android.carrent.utils.extensions.*
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.GeoPoint
 import kotlinx.android.synthetic.main.fragment_register.*
 import kotlinx.android.synthetic.main.fragment_register.view.*
