+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".fragments.home.HomeFragment">
+        tools:context=".fragments.home.HomeFragment.HomeFragment">
 
     <LinearLayout
             xmlns:android="http://schemas.android.com/apk/res/android"
